feat: FA - Pack struct into one slot to save another SSTORE
notes:
- As the maximum amount of the total claim of the merkle root is
known to us, we know it is less than the max value of uint128
therefore we can pack the struct storing the user info into one slot.

You can see the additional gas savings below.

testOnlyOwnerCanSweepTokensinFirstWeek() (gas: -22 (-0.000%))
testOnlyOwnerCanSweepTokensAfterDeadline() (gas: -22 (-0.000%))
testCanClaimMerkleDrop() (gas: -22 (-0.001%))
testOnlyOwnerCanSweepToken() (gas: -22 (-0.001%))
testOwnerCannotSweepTokensBeforeDeadline() (gas: -22 (-0.001%))
testCanNotSeeVestedAmountWithIncorrectAddress() (gas: 107 (0.003%))
testCanSeeVestedAmount() (gas: 112 (0.003%))
testUserCannotCallClaimWithoutInitialClaimFirst() (gas: -2062 (-0.112%))
testUserCanClaimInitialAmountAndOnGoingClaim() (gas: -21664 (-0.151%))
testUserCannotClaimMoreThenTotalClaim() (gas: -21642 (-0.152%))
testUserCannotDoInitialClaimTwice() (gas: -21995 (-0.169%))
testCanDoInitialClaim() (gas: -22024 (-0.180%))
testUserCanClaimFullAmountOnInitialClaim() (gas: -22024 (-0.182%))
Overall gas change: -111302 (-0.945%)

+++ src/GMerkleVestor.sol
@@ -28,9 +28,10 @@ contract GMerkleVestor is Ownable {
                     STORAGE VARIABLES & TYPES
     //////////////////////////////////////////////////////////////*/
 
+	// we can pack into one slot as totalClaim is less max type(uint128).max
 	struct UserInfo {
-		uint256 totalClaim;
-		uint256 claimedAmount;
+		uint128 totalClaim;
+		uint128 claimedAmount;
 	}
 
 	uint256 internal constant ONE_MONTH_SECONDS = 2629746; // average year (including leap years) in seconds / 12
@@ -161,7 +162,8 @@ contract GMerkleVestor is Ownable {
 		}
 
 		// update usersInfo mapping
-		usersInfo[msg.sender] = UserInfo(amount, currentClaimableAmount);
+		// we can downcast safely as the max value of account is less than type(uint128).max
+		usersInfo[msg.sender] = UserInfo(uint128(amount), uint128(currentClaimableAmount));
 
 		// transfer funds to user
 		IERC20(token).safeTransfer(msg.sender, currentClaimableAmount);
@@ -190,8 +192,8 @@ contract GMerkleVestor is Ownable {
 
 		// update claimed amount for user in storage
 		usersInfo[msg.sender].claimedAmount =
-			currentPosition.claimedAmount +
-			currentClaimableAmount;
+			uint128(currentPosition.claimedAmount) +
+			uint128(currentClaimableAmount);
 
 		// transfer funds to user
 		IERC20(token).safeTransfer(msg.sender, currentClaimableAmount);
